[FE] REFACTOR: 페이지 구조, 불필요한 컴포넌트 삭제 #982
